requirements.txt
streamlit
pandas
requests
matplotlib
google-generativeai
tensorflow>=2.16.1
numpy